python/requirements.txt
protobuf==6.33.0